requirements.txt
asgiref==3.4.1
Django==3.2.5
django-taggit==1.5.1
numpy==1.21.1
pandas==1.3.1
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
